requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
aiofiles=23.2.1=pypi_0
aiohttp=3.9.5=pypi_0
aiosignal=1.3.1=pypi_0
altair=5.3.0=pypi_0
annotated-types=0.7.0=pypi_0
anyio=4.4.0=pypi_0
attrs=23.2.0=pypi_0
blinker=1.8.2=pypi_0
branca=0.7.2=pypi_0
bzip2=1.0.8=h80987f9_6
ca-certificates=2024.7.2=hca03da5_0
certifi=2024.7.4=pypi_0
charset-normalizer=3.3.2=pypi_0
click=8.1.7=pypi_0
contourpy=1.2.1=pypi_0
cycler=0.12.1=pypi_0
distro=1.9.0=pypi_0
dnspython=2.6.1=pypi_0
email-validator=2.2.0=pypi_0
fastapi=0.111.1=pypi_0
fastapi-cli=0.0.4=pypi_0
ffmpy=0.3.2=pypi_0
filelock=3.15.4=pypi_0
flask=3.0.3=pypi_0
folium=0.17.0=pypi_0
fonttools=4.53.1=pypi_0
frozenlist=1.4.1=pypi_0
fsspec=2024.6.1=pypi_0
googlemaps=4.10.0=pypi_0
gradio=4.38.1=pypi_0
gradio-client=1.1.0=pypi_0
h11=0.14.0=pypi_0
httpcore=1.0.5=pypi_0
httptools=0.6.1=pypi_0
httpx=0.27.0=pypi_0
huggingface-hub=0.24.0=pypi_0
idna=3.7=pypi_0
importlib-resources=6.4.0=pypi_0
itsdangerous=2.2.0=pypi_0
jinja2=3.1.4=pypi_0
jsonpatch=1.33=pypi_0
jsonpointer=3.0.0=pypi_0
jsonschema=4.23.0=pypi_0
jsonschema-specifications=2023.12.1=pypi_0
kiwisolver=1.4.5=pypi_0
langchain=0.2.10=pypi_0
langchain-core=0.2.22=pypi_0
langchain-openai=0.1.17=pypi_0
langchain-text-splitters=0.2.2=pypi_0
langsmith=0.1.93=pypi_0
libffi=3.4.4=hca03da5_1
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.5=pypi_0
matplotlib=3.9.1=pypi_0
mdurl=0.1.2=pypi_0
multidict=6.0.5=pypi_0
ncurses=6.4=h313beb8_0
numpy=1.26.4=pypi_0
openai=1.36.0=pypi_0
openssl=1.1.1w=h1a28f6b_0
orjson=3.10.6=pypi_0
packaging=24.1=pypi_0
pandas=2.2.2=pypi_0
pillow=10.4.0=pypi_0
pip=24.0=py311hca03da5_0
polyline=2.0.2=pypi_0
pydantic=2.8.2=pypi_0
pydantic-core=2.20.1=pypi_0
pydub=0.25.1=pypi_0
pygments=2.18.0=pypi_0
pyparsing=3.1.2=pypi_0
python=3.11.0=hc0d8a6c_3
python-dateutil=2.9.0.post0=pypi_0
python-dotenv=1.0.1=pypi_0
python-multipart=0.0.9=pypi_0
pytz=2024.1=pypi_0
pyyaml=6.0.1=pypi_0
readline=8.2=h1a28f6b_0
referencing=0.35.1=pypi_0
regex=2024.5.15=pypi_0
requests=2.32.3=pypi_0
rich=13.7.1=pypi_0
rpds-py=0.19.0=pypi_0
ruff=0.5.3=pypi_0
semantic-version=2.10.0=pypi_0
setuptools=69.5.1=py311hca03da5_0
shellingham=1.5.4=pypi_0
six=1.16.0=pypi_0
sniffio=1.3.1=pypi_0
sqlalchemy=2.0.31=pypi_0
sqlite=3.45.3=h80987f9_0
starlette=0.37.2=pypi_0
tenacity=8.5.0=pypi_0
tiktoken=0.7.0=pypi_0
tk=8.6.14=h6ba3021_0
tomlkit=0.12.0=pypi_0
toolz=0.12.1=pypi_0
tqdm=4.66.4=pypi_0
typer=0.12.3=pypi_0
typing-extensions=4.12.2=pypi_0
tzdata=2024.1=pypi_0
urllib3=2.2.2=pypi_0
uvicorn=0.30.3=pypi_0
uvloop=0.19.0=pypi_0
watchfiles=0.22.0=pypi_0
websockets=11.0.3=pypi_0
werkzeug=3.0.3=pypi_0
wheel=0.43.0=py311hca03da5_0
xyzservices=2024.6.0=pypi_0
xz=5.4.6=h80987f9_1
yarl=1.9.4=pypi_0
zlib=1.2.13=h18a0788_1
